Move `dist.py -> megatron/dist.py`
